Move mask-based SQL dispatch from terminal emitters to carrier class (#78)
* Move mask-based SQL dispatch from terminal emitters to carrier class (#77)

Centralizes SQL ownership on the carrier class where the mask and parameters
already live. Single-variant chains emit a static readonly string _sql field;
multi-variant chains emit a static readonly string[] _sql array indexed by
mask value, replacing the per-terminal switch expression with a direct array
lookup. Diagnostics now reference the carrier's _sql field for SqlVariantDiagnostic
entries instead of re-declaring SQL strings inline.


* Narrow mask types from ulong to int and use null! for array gaps

SqlVariants keys and PossibleMasks were ulong but max mask value is 255
(8-bit cap). Narrowing to int removes unnecessary casts and UL suffixes.
Gap entries in multi-variant _sql arrays now emit null! instead of ""
so a mask routing bug surfaces as NRE rather than silently running empty SQL.


* Narrow diagnostics mask types from ulong to int

QueryDiagnostics.ActiveMask, SqlVariants dictionary key, and generated
__variants dictionary all narrowed from ulong to int to match the
generator-side narrowing.


---------

+++ src/Quarry.Generator/QuarryGenerator.cs
@@ -716,7 +716,7 @@ public sealed class QuarryGenerator : IIncrementalGenerator
     /// The carrier terminal expands these tokens at runtime based on the actual collection size.
     /// </summary>
     private static void TokenizeCollectionParameters(
-        Dictionary<ulong, IR.AssembledSqlVariant> sqlMap,
+        Dictionary<int, IR.AssembledSqlVariant> sqlMap,
         IReadOnlyList<IR.QueryParameter> chainParams,
         SqlDialect dialect)
     {
@@ -732,7 +732,7 @@ public sealed class QuarryGenerator : IIncrementalGenerator
 
         // Replace placeholders with tokens in all SQL variants.
         // Collect updates and apply after iteration to avoid allocating a key list copy.
-        var pendingUpdates = new List<(ulong Key, string Sql, int ParamCount)>();
+        var pendingUpdates = new List<(int Key, string Sql, int ParamCount)>();
         foreach (var kvp in sqlMap)
         {
             var sql = kvp.Value.Sql;
